backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic

chromadb
sentence-transformers
torch

openai
python-dotenv
langfuse

pymupdf
pdfplumber